backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-environ
django-filter
django-redis